Silencing Implicit Conversion from Nullable to Non-Null Pointers Error
Silencing the error caused by implicitly using a nullable pointer in a non-null context. Simply storing the value into a variable first silences the warning but this does not fix the underlying, potential, problem. The numbers can still be null, which will cause an exception when accessed in an array.