Xcode 9 Updated Settings
Code signing was disabled as this is not required for at the framework level. Frameworks are embedded into applications so only the final app needs to be signed.

Enabled new warnings to help catch some common and obscure errors such as infinite recursion, autorelease objects being captured by blocks, etc.